==> Auto_Codes/2_Data_Creation.ipynb <==
!pip install orjson

import pandas as pd
import numpy as np
import json
import os
import orjson
import glob
import sys
import time
from collections import defaultdict
from datetime import datetime,timedelta
import math
from collections import Counter
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_rows',None)
pd.set_option('display.max_columns',None)
pd.set_option('display.max_colwidth', None)

### Reading Data
df= pd.read_csv('s3://scien-ds-sm-root-backup/home/sushil/LN_Unified_V2/Client_Data_Auto/210324_LN_V2_Auto_Dataset_With_Performance_Race_Age_and_Gender.csv')
df.shape

df[['approve_flag', 'funded_flag', 'bad_flag']].sum()

df['client'].value_counts(dropna= False)

df= df[df['client']!= 'AAC'].reset_index(drop= True)
df.shape

df= df[df['auto_score']>500].reset_index(drop= True)
df.shape

df['app_date']= df['app_date'].apply(lambda x: pd.to_datetime(x[:10]) if pd.isna(x)== False else x)
df['app_date'].describe()

def get_age(app_date, dob):
#     app_date= app_date.copy()
#     dob= dob.copy()
    
    if (pd.isna(app_date)== False) and (pd.isna(dob)== False):
        
        app_date= pd.to_datetime(app_date)
        dob= pd.to_datetime(dob)
        
        if dob >= app_date:
            dob= dob.replace(year= dob.year - 100)
        else:
            dob= dob

        app_year, app_month, app_day= app_date.year, app_date.month, app_date.day
        dob_year, dob_month, dob_day= dob.year, dob.month, dob.day
        
        age= (app_year - dob_year) + ((app_month - dob_month)/12) + ((app_day - dob_day)/365.5)
        
        return age
    
    else:
        return np.NaN

df[df['applicant_age']<0].shape

df[df['applicant_age']<0][['client', 'applicant_age', 'applicant_dob', 'app_date']].sample(10)

df['applicant_age']= df.apply(lambda x: get_age(x['app_date'], x['applicant_dob']) if x['applicant_age']< 0 else x['applicant_age'], axis= 1)
df['applicant_age'].describe()

df.loc[732066][['client', 'applicant_age', 'applicant_dob', 'app_date']]

# df.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_All_Updated_Data_220324.csv', index= False)

df= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_All_Updated_Data_220324.csv')
df.shape

from sklearn.model_selection import train_test_split, StratifiedShuffleSplit, StratifiedGroupKFold, StratifiedKFold

# sgkf= StratifiedGroupKFold(n_splits= 5, shuffle= True, random_state= 127)

T1,T2= train_test_split(df, test_size= 0.15, random_state= 2022)
T1.shape, T2.shape

T1[['approve_flag', 'funded_flag', 'bad_flag']].sum(), T2[['approve_flag', 'funded_flag', 'bad_flag']].sum()

T1['bad_flag'].sum()/T1['funded_flag'].sum(), T2['bad_flag'].sum()/T2['funded_flag'].sum()

T1['approve_flag'].mean(), T2['approve_flag'].mean()

T1.groupby('client').agg(
    ttd= ('unique_id', 'count'),
    approved= ('approve_flag', 'sum'),
    funded= ('funded_flag', 'sum'),
    bad= ('bad_flag', 'sum'),
    avg_fico= ('fico_pb', 'mean'),
    avg_auto_score= ('auto_score', 'mean'),
    min_app_date=  ('app_date', 'min'),
    max_app_date= ('app_date', 'max')
)

T2.groupby('client').agg(
    ttd= ('unique_id', 'count'),
    approved= ('approve_flag', 'sum'),
    funded= ('funded_flag', 'sum'),
    bad= ('bad_flag', 'sum'),
    avg_fico= ('fico_pb', 'mean'),
    avg_auto_score= ('auto_score', 'mean'),
    min_app_date=  ('app_date', 'min'),
    max_app_date= ('app_date', 'max')
)

# T2.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_OOT_Data_220324.csv', index= False)
# T1.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_DEV_Data_220324.csv', index= False)

del df, T2

### Now Splitting the data into Train and Test

T1.shape

T1['Borrower_Race_Prim_BISG'].value_counts(dropna= False)

T1['borrower_gender'].value_counts(dropna= False)

T1['Race']= T1['Borrower_Race_Prim_BISG']
T1['Gender']= T1['borrower_gender']
T1['age_cuts']= np.where(T1['applicant_age'].isnull(), 'missing',
                         np.where(T1['applicant_age']< 62, 'Age below 62',
                                  np.where(T1['applicant_age']>= 62, 'Age 62 and over', 'missing')))

T1['age_cuts'].value_counts(dropna= False)

T1.groupby('age_cuts')['applicant_age'].agg(['min', 'max'])

T3,T4= train_test_split(T1, test_size= 0.3, random_state= 139)
T3.shape, T4.shape

T3[['approve_flag', 'funded_flag', 'bad_flag']].sum(), T4[['approve_flag', 'funded_flag', 'bad_flag']].sum()

T3['approve_flag'].mean(), T4['approve_flag'].mean()

# T3.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_Train_RI_Data_220324.csv', index= False)
# T4.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_Test_RI_Data_220324.csv', index= False)

### Next Part Funded

Train = pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_Train_RI_Data_220324.csv')
Test = pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_Test_RI_Data_220324.csv')
Oot= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_V2_OOT_Data_220324.csv')

Train.shape, Test.shape, Oot.shape

Oot['Race']= Oot['Borrower_Race_Prim_BISG']
Oot['Gender']= Oot['borrower_gender']
Oot['age_cuts']= np.where(Oot['applicant_age'].isnull(), 'missing',
                         np.where(Oot['applicant_age']< 62, 'Age below 62',
                                  np.where(Oot['applicant_age']>= 62, 'Age 62 and over', 'missing')))

Oot['age_cuts'].value_counts(dropna= False)

TTD= pd.concat([Train, Test]).reset_index(drop= True)
TTD.shape

funded_dev= TTD[TTD['funded_flag']== 1].reset_index(drop= True)
funded_dev.shape

scored= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/RI_Auto_Model/250324_RI_Auto_Model_Scored_Data.csv')
scored.shape

scored['unique_id']= scored['unique_id'].astype(str)
funded_dev['unique_id']= funded_dev['unique_id'].astype(str)

funded_dev= funded_dev.merge(scored[['unique_id', 'client', 'p1', 'weight']], on= ['unique_id', 'client'], how= 'left')
funded_dev.shape

funded_dev[['p1', 'weight']].isnull().sum()

funded_dev.rename(columns= {'p1': 'p1_ri'}, inplace= True)

funded_dev[['p1_ri', 'weight']].describe()

from sklearn.model_selection import train_test_split

F1, F2= train_test_split(funded_dev, test_size= 0.3, random_state= 91)
print(F1['bad_flag'].mean(), F2['bad_flag'].mean())
print(sum(F1['bad_flag']*F1['weight'])/F1['weight'].sum(), sum(F2['bad_flag']*F2['weight'])/F2['weight'].sum())

F1.shape, F2.shape

# F1.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/2_LN_Auto_V2_Train_Funded_Data_250324.csv', index= False)
# F2.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/2_LN_Auto_V2_Test_Funded_Data_250324.csv', index= False)
# Oot[Oot['funded_flag']== 1].to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/2_LN_Auto_V2_Oot_Funded_Data_250324.csv', index= False)

Oot[Oot['funded_flag']== 1].shape

# TTD.to_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/2_LN_Auto_V2_TTD_Data_250324.csv', index= False)
