==> Auto_Codes/4.3_Auto_LN_V2_Protected_Class.ipynb <==
!pip install xgboost

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import json
import logging
import xgboost as xgb
import ast
import itertools
import scipy.stats as ss
from sklearn.preprocessing import MinMaxScaler
from scipy.stats import chi2
import os

!pip install hyperopt

import numpy as np
import pandas as pd
from hyperopt import Trials, STATUS_OK, tpe, hp, fmin
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 200)
import h2o
from h2o.estimators import H2OXGBoostEstimator
from hyperopt.pyll import scope as ho_scope
from hyperopt.pyll.stochastic import sample as ho_sample

import sys
sys.path.append('/root/Unified_LN_V2/')
from Utilities import *

### Starting H2o Session
h2o.init()

#Train = pd.read_csv("Train.csv")
#Test = pd.read_csv("Test.csv")
TTD= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/Mono_Data/3_Auto_Ttd_Binned_Data_260324.csv')
TTD.shape
# oot_binned = pd.read_csv("/home/username/aniruddha/NewBinned_BureauLN_BP_oot_mono_60_18m.csv")

TTD['Race'].value_counts()

TTD["Race_black"] = np.where(TTD["Race"] == "black",1, 0)
TTD["Race_hispanic"] = np.where(TTD["Race"] == "hispanic",1, 0)
TTD["Race_api"] = np.where((TTD["Race"] == "asian")|(TTD["Race"] == "api"),1, 0)
TTD["age_cuts_flag"] = np.where(TTD["age_cuts"] == "Age 62 and over",1, 0)
TTD["Gender_flag"] = np.where(TTD["Gender"] == "F",1, 0)

TTD["Race_api"].value_counts()

TTD['client'].value_counts(dropna= False)

### Convert the Variable into categorical
var_to_cat= '''businesstitleleadership
proflictypecategory
addrcurrentdeedmailing
addrcurrentphoneservice
addrcurrentsubjectowned
addrinputdeedmailing
addrinputphoneservice
addrinputproblems
addrinputsubjectowned
addrlastmoveecontrajectory
addronfilecollege
addronfilecorrectional
addronfilehighrisk
addrprevioussubjectowned
alertregulatorycondition
assetownership
assetpersonal
assetprop
assetpropnewestmortgagetype
bankruptcychapter
bankruptcystatus
businessassociation
educationattendance
educationevidence
educationinstitutionrating
educationprogramattended
inquiryauto12month
inquirybanking12month
inquirycollections12month
inquirynonshortterm12month
inquiryshortterm12month
inquirytelcom12month
shorttermloanrequest
shorttermloanrequest12month
shorttermloanrequest24month
subjectnewestrecord12month
sourcevoterregistration'''.splitlines()
len(var_to_cat)

for var in var_to_cat:
    TTD[var]= TTD[var].astype(str)

TTD.columns

from sklearn.model_selection import train_test_split

TTD_train,TTD_test = train_test_split(TTD, test_size=0.10,random_state=4127)

TTD_train['approve_flag'].mean(), TTD_test['approve_flag'].mean()

req_cols= ['assetpropevercount', 'sourcenonderogcount06month', 'addronfilecount', 'addrinputcountyratio', 'assetpropnewestsaleprice', 'addrcurrenttaxvalue', 'addrinputavmvalue', 'addrinputavmvalue60month', 'addrinputlastsaleprice', 'addrinputtaxmarketvalue', 'sourcecredheadertimeoldest', 'addrinputphonecount', 'assetpropcurrenttaxtotal', 'addrcurrentavmvalue', 'assetproppurchasetimeoldest', 'addrpreviouslengthofres', 'addrinputlengthofres', 'addrcurrenttaxmarketvalue', 'addrprevioustimenewest', 'addrcurrenttractratio', 'sourcenonderogcount12month', 'assetpropcurrentcount', 'addrinputsubjectcount', 'sourcenonderogcount', 'addrinputtaxvalue', 'assetpropeversoldcount', 'businessassociationtimeoldest', 'addrinputtractratio', 'derogtimenewest', 'addrchangecount24month', 'addrinputtimelastsale', 'educationinstitutionprivate', 'addrinputblockratio', 'businesstitleleadership', 'addrcurrentlengthofres', 'addrcurrentavmratio12monthprior', 'addrcurrentblockratio', 'derogcount', 'profliccount', 'addrinputavmratio60monthprior', 'addrcurrentcountyratio', 'assetpropsaletimeoldest', 'subjectrecordtimeoldest', 'addrcurrenttimenewest', 'proflictypecategory', 'criminalnonfelonytimenewest', 'lienjudgmentdollartotal', 'addrcurrentlastsalesprice', 'addrinputtimenewest', 'evictiontimenewest', 'bankruptcycount', 'addrcurrenttimelastsale', 'addrinputtimeoldest', 'addrlastmovetaxratiodiff', 'criminalnonfelonycount', 'sourcenonderogcount03month', 'addrchangecount03month', 'addrchangecount06month', 'addrchangecount12month', 'addrchangecount60month', 'addrcurrentavmratio60monthprior', 'addrcurrentavmvalue12month', 'addrcurrentavmvalue60month', 'addrcurrentcorrectional', 'addrcurrentdeedmailing', 'addrcurrentphoneservice', 'addrcurrentsubjectowned', 'addrcurrenttimeoldest', 'addrinputavmratio12monthprior', 'addrinputavmvalue12month', 'addrinputdeedmailing', 'addrinputphoneservice', 'addrinputproblems', 'addrinputsubjectowned', 'addrlastmoveecontrajectory', 'addronfilecollege', 'addronfilecorrectional', 'addronfilehighrisk', 'addrpreviouscorrectional', 'addrprevioussubjectowned', 'addrprevioustimeoldest', 'alertregulatorycondition', 'assetownership', 'assetpersonal', 'assetpersonalcount', 'assetprop', 'assetpropnewestmortgagetype', 'assetproppurchasecount12month', 'assetproppurchasetimenewest', 'assetpropsalepurchaseratio', 'assetpropsaletimenewest', 'assetpropsoldcount12month', 'bankruptcychapter', 'bankruptcycount24month', 'bankruptcydismissed24month', 'bankruptcystatus', 'bankruptcytimenewest', 'businessassociation', 'criminalfelonycount', 'criminalfelonycount12month', 'criminalfelonytimenewest', 'criminalnonfelonycount12month', 'derogcount12month', 'educationattendance', 'educationevidence', 'educationinstitutionrating', 'educationprogramattended', 'evictioncount', 'evictioncount12month', 'inquiryauto12month', 'inquirybanking12month', 'inquirycollections12month', 'inquirynonshortterm12month', 'inquiryshortterm12month', 'inquirytelcom12month', 'lienjudgmentcount', 'lienjudgmentcount12month', 'lienjudgmentcourtcount', 'lienjudgmentforeclosurecount', 'lienjudgmentothercount', 'lienjudgmentsmallclaimscount', 'lienjudgmenttaxcount', 'lienjudgmenttimenewest', 'shorttermloanrequest', 'shorttermloanrequest12month', 'shorttermloanrequest24month', 'sourcecredheadertimenewest', 'sourcevoterregistration', 'subjectnewestrecord12month', 'subjectrecordtimenewest']
len(req_cols)

TTD_train.shape, TTD_test.shape

TTD['bankruptcystatus'].value_counts(dropna= False)

TTD['bankruptcystatus'].dtypes

ttd_train_h2o = h2o.H2OFrame(TTD_train)
ttd_test_h2o = h2o.H2OFrame(TTD_test)
print(ttd_train_h2o.shape,ttd_test_h2o.shape)

ttd_train_h2o['bankruptcystatus'].unique()

ttd_train_h2o['bankruptcystatus'].asfactor().unique()

for col in var_to_cat:
    ttd_train_h2o[col] = ttd_train_h2o[col].asfactor()
    ttd_test_h2o[col] = ttd_test_h2o[col].asfactor()

for k in TTD['bankruptcychapter'].unique():
    print(k, ttd_train_h2o[ttd_train_h2o['bankruptcychapter']== k].shape)

path_to_save= '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/Protected_Class_Data/'

def recall_cal(data_h2o,data,name,dep):
    predict = xgbm_FL.predict(data_h2o)
    p =predict.as_data_frame()
    dd = data.loc[:,unique_cols + [dep_var]]
    dd = pd.DataFrame(dd)
    dd = dd.reset_index()
    dd['RECWGT'] = 1
    dd['conwt'] = 1
    pred = pd.concat([dd.loc[:,(dep_var,"RECWGT","conwt")],p],axis=1)
    dep_flag_rate = (pred["RECWGT"]*pred[dep_var]).sum()/pred["RECWGT"].sum()
    pred = pred.sort_values(by=["p1"],ascending=False).reset_index()
    pred["cum_wgt"] = pred["RECWGT"].cumsum()
    pred["cum_wgt_nor"] = pred["cum_wgt"]/pred["RECWGT"].sum()
    pred["scorecut"] = np.where(pred["cum_wgt_nor"]<=dep_flag_rate,1,0)
    pred["TP"] = np.where((pred["scorecut"]==1) & (pred[dep_var]==1),1,0)
    recall = pred["TP"].sum()/pred[dep_var].sum()
    print( pred["TP"].sum(),pred[dep_var].sum())
    pred.to_csv(path_to_save+name+dep_var +"_dataset.csv")
    return recall, dep_flag_rate

unique_cols = ['unique_id', 'client']
wt_col = "wt2"
folds = 10
predictors = ttd_train_h2o.columns
req_cols = req_cols# config.dropna().Feature_Name.to_list()
model_cols = unique_cols + req_cols
ttd_train_h2o=ttd_train_h2o[model_cols+['Race_black','Race_hispanic','Race_api','age_cuts_flag','Gender_flag']]#,'mmct_tag']]


print(ttd_train_h2o.shape)
print(len(req_cols))

%%time
#dep_var = "Race_black"
data_recall = pd.DataFrame()
for dep_var in ['Race_black','Race_hispanic','Race_api','age_cuts_flag','Gender_flag']:
    print(dep_var)

    xgbm_FL = H2OXGBoostEstimator(
        ntrees=2000,    learn_rate=0.01,
        max_depth=3,    model_id=dep_var,
        distribution = "bernoulli",    seed=42,
        nfolds = 0,    ignored_columns = unique_cols)
    model_cols = unique_cols + [dep_var]+ req_cols
    ttd_train_h2o[dep_var] = ttd_train_h2o[dep_var].asfactor()
    xgbm_FL.train(y=dep_var, training_frame=ttd_train_h2o[model_cols])
    xgbm_FL.download_mojo(path = path_to_save+dep_var+".zip")
    recall,dep_flag_rate = recall_cal(ttd_train_h2o,TTD_train,"Train",dep_var)
    print(recall, dep_flag_rate)
    recall1,dep_flag_rate1 = recall_cal(ttd_test_h2o,TTD_test,"Test",dep_var)
    print(recall1,dep_flag_rate1)
    var_imp = xgbm_FL._model_json['output']['variable_importances'].as_data_frame()
    var_imp['Var_unique'] = var_imp['variable'].str.split('.', expand=True)[0]
    var_imp.to_csv(path_to_save+dep_var +"_var_im.csv")
    df = pd.DataFrame([[dep_var,recall,recall1,dep_flag_rate,dep_flag_rate1]], columns=['Dep', 'recall',"test_recall","dep_flag_rate","dep_flag_rate1"])
    data_recall = pd.concat([data_recall, df])#data_recall.append(df)
    data_recall.to_csv(path_to_save +"recall.csv")

df

data_recall

# predict = xgbm_black.predict(ttd_test_h2o)
# p =predict.as_data_frame()
# dd = TTD_test.loc[:,unique_cols + [dep_var]]
# dd = pd.DataFrame(dd)
# dd = dd.reset_index()
# dd['RECWGT'] = 1
# dd['conwt'] = 1
# pred = pd.concat([dd.loc[:,(dep_var,"RECWGT","conwt")],p],axis=1)
# print(pred["RECWGT"].sum(),((pred["RECWGT"]*pred[dep_var]).sum()),pred["RECWGT"].max())
# dep_flag_rate = (pred["RECWGT"]*pred[dep_var]).sum()/pred["RECWGT"].sum()
# pred = pred.sort_values(by=["p1"],ascending=False).reset_index()
# pred["cum_wgt"] = pred["RECWGT"].cumsum()
# pred["cum_wgt_nor"] = pred["cum_wgt"]/pred["RECWGT"].sum()
# pred["scorecut"] = np.where(pred["cum_wgt_nor"]<=dep_flag_rate,1,0)
# pred["TP"] = np.where((pred["scorecut"]==1) & (pred[dep_var]==1),1,0)
# recall_test = pred["TP"].sum()/pred[dep_var].sum()
# print(recall_test) 

# var_imp = xgbm_black._model_json['output']['variable_importances'].as_data_frame()
# var_imp['Var_unique'] = var_imp['variable'].str.split('.', expand=True)[0]
# var_imp.head(2)

# var_imp.to_csv(path_to_save+dep_var +"_var_im.csv")

data_recall= pd.read_csv(path_to_save +"recall.csv")
data_recall

%%time
#dep_var = "Race_black"
# data_recall = pd.DataFrame()
for dep_var in ['Gender_flag']:#['Race_black','Race_hispanic','Race_api','age_cuts_flag','Gender_flag']:
    print(dep_var)

    xgbm_FL = H2OXGBoostEstimator(
        ntrees=2000,    learn_rate=0.01,
        max_depth=3,    model_id=dep_var,
        distribution = "bernoulli",    seed=42,
        nfolds = 0,    ignored_columns = unique_cols)
    model_cols = unique_cols + [dep_var]+ req_cols
    ttd_train_h2o[dep_var] = ttd_train_h2o[dep_var].asfactor()
    xgbm_FL.train(y=dep_var, training_frame=ttd_train_h2o[model_cols])
    xgbm_FL.download_mojo(path = path_to_save+dep_var+".zip")
    recall,dep_flag_rate = recall_cal(ttd_train_h2o,TTD_train,"Train",dep_var)
    print(recall, dep_flag_rate)
    recall1,dep_flag_rate1 = recall_cal(ttd_test_h2o,TTD_test,"Test",dep_var)
    print(recall1,dep_flag_rate1)
    var_imp = xgbm_FL._model_json['output']['variable_importances'].as_data_frame()
    var_imp['Var_unique'] = var_imp['variable'].str.split('.', expand=True)[0]
    var_imp.to_csv(path_to_save+dep_var +"_var_im.csv")
    df = pd.DataFrame([[dep_var,recall,recall1,dep_flag_rate,dep_flag_rate1]], columns=['Dep', 'recall',"test_recall","dep_flag_rate","dep_flag_rate1"])
    data_recall = pd.concat([data_recall, df])#data_recall.append(df)
    data_recall.to_csv(path_to_save +"recall.csv")

data_recall

TTD['Race_black'].sum(), len(TTD)

TTD['Race_hispanic'].sum(), len(TTD)

TTD['Race_api'].sum(), len(TTD)

TTD['age_cuts_flag'].sum(), len(TTD)

TTD['Gender_flag'].sum(), len(TTD)

train= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/Mono_Data/3_Auto_Train_Binned_Data_260324.csv')
test= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/Mono_Data/3_Auto_Test_Binned_Data_260324.csv')
oot= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/Mono_Data/3_Auto_Oot_Binned_Data_260324.csv')
ttd= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/Mono_Data/3_Auto_Ttd_Binned_Data_260324.csv')
train.shape, test.shape, oot.shape

ttd.shape

train.groupby('client')['fico_pb'].apply(lambda x: x.isnull().sum())

test.groupby('client')['fico_pb'].apply(lambda x: x.isnull().sum())

oot.groupby('client')['fico_pb'].apply(lambda x: x.isnull().sum())

ttd.groupby('client')['fico_pb'].apply(lambda x: x.isnull().sum())

ttd[ttd['funded_flag']==1].groupby('client')['fico_pb'].apply(lambda x: x.isnull().sum())

pd.Series(['addrinputcountyratio', "addrinputavmvalue12month","addrinputavmvalue60month","addrcurrentavmvalue12month","addrinputavmvalue","sourcevoterregistration","addrinputavmratio60monthprior","addrcurrentcountyratio","addrcurrentphoneservice","addrinputsubjectcount","proflictypecategory","assetpersonal","educationevidence"]).reset_index()
