==> Auto_Codes/7.1_Model_Comparison Asset Ownership.ipynb <==
### Final Results With Comparison to Different Model Performance

!pip install xgboost

import warnings
warnings.filterwarnings('ignore')
import pandas as pd
print(pd.__version__)
import numpy as np
import json
import logging
import xgboost as xgb
import ast
import itertools
import scipy.stats as ss
from sklearn.preprocessing import MinMaxScaler
from scipy.stats import chi2
import os
import sys

pd.set_option('display.max_rows',None)
pd.set_option('display.max_columns',None)
pd.set_option('display.max_colwidth', None)

sys.path.append('/root/Unified_LN_V2/')
from Utilities import *

def air_calc_new(pred,prefix):
    air = {}
    target_control_dict = [
            {'age_cuts': {
                'target': ['Age 62 and over'],
                'control': ['Age below 62']
            }},
            {'Race': {
                'target': ['black', 'hispanic'],
                'control': ['white']
            }},
            {'Race': {
                'target': ['black'],
                'control': ['white']
            }},
            {'Race': {
                'target': ['hispanic'],
                'control': ['white']
            }},
            {'Gender': {
                'target': ['F'],
                'control': ['M']
            }},
            # 'mmct_tag': {
            #     'target': ['mmct'],
            #     'control': ['non-mmct']
            # }
    ]

    for var in target_control_dict:
        for i in var:
            target_list = var[i]['target']
            control_list = var[i]['control']
            p_target_list= ','.join(target_list) if (
                (len(target_list)<= 1) and (('black' in target_list) or ('hispanic' in target_list))
            ) else ''
            
            ttd_target= pred[pred[i].isin(target_list)]['con_wt'].sum()
            approved_target= pred[pred[i].isin(target_list)]['approved'].multiply(pred[pred[i].isin(target_list)]['con_wt']).sum()
            ttd_control= pred[pred[i].isin(control_list)]['con_wt'].sum()
            approved_control= pred[pred[i].isin(control_list)]['approved'].multiply(pred[pred[i].isin(control_list)]['con_wt']).sum()
            app_pct_target= round(approved_target/ttd_target, 4)
            app_pct_control= round(approved_control/ttd_control, 4)
            air[prefix+i+p_target_list]= [round(app_pct_target/app_pct_control, 4)]
        
            # print({'variable': i, 'target_list': p_target_list, 'ttd_target': ttd_target, 'ttd_control': ttd_control, 
                   # 'approved_target': approved_target, 'approved_control': approved_control})
            
    # print(air)
        
    return air

col_tuple_pass= ["unique_id", 'client', "wt2", 'fico_pb', 'auto_score', 'approve_flag', 'funded_flag', 'bad_flag', 'Race', 'Gender', 'age_cuts']

def get_risk_rank(data, target, cols_tuple, score_col= 'p1', weight= 'RECWGT', asc= False):
    dd = data.loc[:, cols_tuple + [score_col]]
    # dd = pd.DataFrame(dd)
    # dd = dd.reset_index()
    dd['RECWGT'] = dd["wt2"]
    dd['conwt'] = 1
    dd[target] = dd[target] 
    # pred = pd.concat([dd.loc[:, cols_tuple + ['RECWGT', 'conwt']],p],axis=1)
    pred= dd.copy()
    # display(pred.head())
    print(pred[weight].sum(),((pred[weight]*pred[target]).sum()),pred[weight].max())

    data_ks_cap5= rnkOrderingN(pred, target, weight, score_col, 5, asc)
    # display(data_ks_cap5[0])
    # print(data_ks_cap5[1])
    # data_ks_cap10= rnkOrderingN(pred, target, weight, score_col, 10, asc)
    # display(data_ks_cap10[0])
    # print(data_ks_cap10[1])
    return data_ks_cap5

def get_segment(d):
    if (d['derogcount']> 0):
        return 'Derog'
    elif ((d['derogcount']== 0) and (d['assetprop']> 0)):
        return 'Property Owner'
    elif ((d['derogcount']== 0) and (d['assetprop']== 0)):
        return 'Non Property Owner'
    else:
        return 'Missing'

def Equal_Approval_Cutoff_Barsath(pred, target, score_col, asc= True):
    pred= pred.copy()
    actual_approved_count= pred[target].sum()
    pred= pred.sort_values(by= [score_col], ascending= asc)
    pred['con_wt']= 1
    pred['cumsum']= pred['con_wt'].cumsum()
    pred['approved']= np.where(pred['cumsum']<= actual_approved_count, 1, 0)
    return pred #cutoff, original_dataset

## Reading Scored Data - Overall LDA

path_overall= '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/With_LDA/Iteration_3/' #/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/With_LDA/
os.listdir(path_overall)

# train_scored_all= pd.read_csv(f"{path_overall}With_LDA_Train_Auto_LN_V2_Scored_Data_2024-03-29.csv")
# test_scored_all= pd.read_csv(f"{path_overall}With_LDA_Test_Auto_LN_V2_Scored_Data_2024-03-29.csv")
# oot_scored_all= pd.read_csv(f"{path_overall}With_LDA_Oot_Auto_LN_V2_Scored_Data_2024-03-29.csv")
# ttd_scored_all= pd.read_csv(f"{path_overall}With_LDA_TTD_Auto_LN_V2_Scored_Data_2024-03-29.csv")

# train_scored_all= pd.read_csv(f"{path_overall}With_LDA_Train_Auto_LN_V2_Scored_Data_2024-04-02.csv")
# test_scored_all= pd.read_csv(f"{path_overall}With_LDA_Test_Auto_LN_V2_Scored_Data_2024-04-02.csv")
# oot_scored_all= pd.read_csv(f"{path_overall}With_LDA_Oot_Auto_LN_V2_Scored_Data_2024-04-02.csv")
# ttd_scored_all= pd.read_csv(f"{path_overall}With_LDA_TTD_Auto_LN_V2_Scored_Data_2024-04-02.csv")

train_scored_all= pd.read_csv(f"{path_overall}With_LDA_Train_Auto_LN_V2_Scored_Data_2024-04-08.csv")
test_scored_all= pd.read_csv(f"{path_overall}With_LDA_Test_Auto_LN_V2_Scored_Data_2024-04-08.csv")
oot_scored_all= pd.read_csv(f"{path_overall}With_LDA_Oot_Auto_LN_V2_Scored_Data_2024-04-08.csv")
ttd_scored_all= pd.read_csv(f"{path_overall}With_LDA_TTD_Auto_LN_V2_Scored_Data_2024-04-08.csv")

train_scored_all.shape, test_scored_all.shape, oot_scored_all.shape, ttd_scored_all.shape

funded_scored_all= pd.concat([train_scored_all, test_scored_all, oot_scored_all]).reset_index(drop= True)
funded_scored_all.shape, ttd_scored_all.shape, ttd_scored_all['funded_flag'].sum()

funded_scored_all['unique_id']= funded_scored_all['unique_id'].astype(str)
ttd_scored_all['unique_id']= ttd_scored_all['unique_id'].astype(str)

funded_scored_all.rename(columns= {'p1': 'p1_overall_lda'}, inplace= True)
ttd_scored_all.rename(columns= {'p1': 'p1_overall_lda'}, inplace= True)

### Reading V 1.5 Scored Data

v15= pd.read_csv('/root/Unified_LN_V2/LN_Auto_V2/Data/040424_Auto_LN_V1_5_Scored_Data.csv')
v15.head()

v15.shape, ttd_scored_all.shape, funded_scored_all.shape

v15['unique_id']= v15['unique_id'].astype(str)

print(ttd_scored_all.shape)
ttd_scored_all= ttd_scored_all.merge(v15[['unique_id', 'client', 'p1_v1_5']], on= ['unique_id', 'client'], how= 'left')
ttd_scored_all.shape

ttd_scored_all['p1_v1_5'].isnull().sum()

ttd_scored_all[['p1_v1_5', 'p1_overall_lda']].describe()

print(funded_scored_all.shape)
funded_scored_all= funded_scored_all.merge(v15[['unique_id', 'client', 'p1_v1_5']], on= ['unique_id', 'client'], how= 'left')
funded_scored_all.shape, funded_scored_all['p1_v1_5'].isnull().sum()

# rnkOrderingN(funded_scored_all[~(funded_scored_all['fico_pb']>= 740)], 'bad_flag', None, 'p1_overall_lda', 5)

# rnkOrderingN(funded_scored_all[~(funded_scored_all['fico_pb']>= 740)], 'bad_flag', None, 'p1_v1_5', 5)

#### Reading Property Owner Dataset

path_owner= '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Property_Owner_Model/'
os.listdir(path_owner)

train_scored_owner= pd.read_csv(f'{path_owner}Property_Owner_Train_Auto_LN_V2_Scored_Data_2024-03-29.csv')
test_scored_owner= pd.read_csv(f'{path_owner}Property_Owner_Test_Auto_LN_V2_Scored_Data_2024-03-29.csv')
oot_scored_owner= pd.read_csv(f'{path_owner}Property_Owner_Oot_Auto_LN_V2_Scored_Data_2024-03-29.csv')
ttd_scored_owner= pd.read_csv(f'{path_owner}Property_Owner_TTD_Auto_LN_V2_Scored_Data_2024-03-29.csv')
train_scored_owner.shape, test_scored_owner.shape, oot_scored_owner.shape, ttd_scored_owner.shape

train_scored_owner['unique_id']= train_scored_owner['unique_id'].astype(str)
test_scored_owner['unique_id']= test_scored_owner['unique_id'].astype(str)
oot_scored_owner['unique_id']= oot_scored_owner['unique_id'].astype(str)
ttd_scored_owner['unique_id']= ttd_scored_owner['unique_id'].astype(str)

train_scored_owner['dataset']= 'Property Owner'
test_scored_owner['dataset']= 'Property Owner'
oot_scored_owner['dataset']= 'Property Owner'
ttd_scored_owner['dataset']= 'Property Owner'

#### Reading Non Property Owner Dataset

path_non_owner= '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Non_Property_Owner_Model/'
os.listdir(path_non_owner)

train_scored_nonowner= pd.read_csv(f'{path_non_owner}Non_Property_Owner_Train_Auto_LN_V2_Scored_Data_2024-03-29.csv')
test_scored_nonowner= pd.read_csv(f'{path_non_owner}Non_Property_Owner_Test_Auto_LN_V2_Scored_Data_2024-03-29.csv')
oot_scored_nonowner= pd.read_csv(f'{path_non_owner}Non_Property_Owner_Oot_Auto_LN_V2_Scored_Data_2024-03-29.csv')
ttd_scored_nonowner= pd.read_csv(f'{path_non_owner}Non_Property_Owner_TTD_Auto_LN_V2_Scored_Data_2024-03-29.csv')
train_scored_nonowner.shape, test_scored_nonowner.shape, oot_scored_nonowner.shape, ttd_scored_nonowner.shape

train_scored_nonowner['unique_id']= train_scored_nonowner['unique_id'].astype(str)
test_scored_nonowner['unique_id']= test_scored_nonowner['unique_id'].astype(str)
oot_scored_nonowner['unique_id']= oot_scored_nonowner['unique_id'].astype(str)
ttd_scored_nonowner['unique_id']= ttd_scored_nonowner['unique_id'].astype(str)

train_scored_nonowner['dataset']= 'Non Property Owner'
test_scored_nonowner['dataset']= 'Non Property Owner'
oot_scored_nonowner['dataset']= 'Non Property Owner'
ttd_scored_nonowner['dataset']= 'Non Property Owner'

train_ono= pd.concat([train_scored_owner, train_scored_nonowner]).reset_index(drop= True)
test_ono= pd.concat([test_scored_owner, test_scored_nonowner]).reset_index(drop= True)
oot_ono= pd.concat([oot_scored_owner, oot_scored_nonowner]).reset_index(drop= True)
ttd_ono= pd.concat([ttd_scored_owner, ttd_scored_nonowner]).reset_index(drop= True)
train_ono.shape, test_ono.shape, oot_ono.shape, ttd_ono.shape

print(train_ono['dataset'].value_counts(dropna= False), train_scored_owner.shape, train_scored_nonowner.shape)
print(test_ono['dataset'].value_counts(dropna= False), test_scored_owner.shape, test_scored_nonowner.shape)
print(oot_ono['dataset'].value_counts(dropna= False), oot_scored_owner.shape, oot_scored_nonowner.shape)
print(ttd_ono['dataset'].value_counts(dropna= False), ttd_scored_owner.shape, ttd_scored_nonowner.shape)

del train_scored_all, train_scored_nonowner, train_scored_owner
del test_scored_all, test_scored_nonowner, test_scored_owner
del oot_scored_all, oot_scored_nonowner, oot_scored_owner
del ttd_scored_nonowner, ttd_scored_owner

import gc
gc.collect()

### Merging p1-overall

print(train_ono.shape)
train_ono= train_ono.merge(funded_scored_all[['unique_id', 'client', 'p1_overall_lda', 'p1_v1_5']], on= ['unique_id', 'client'], how= 'left')
print(train_ono.shape)
print(train_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].isnull().sum())

print(test_ono.shape)
test_ono= test_ono.merge(funded_scored_all[['unique_id', 'client', 'p1_overall_lda', 'p1_v1_5']], on= ['unique_id', 'client'], how= 'left')
print(test_ono.shape)
print(test_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].isnull().sum())

print(oot_ono.shape)
oot_ono= oot_ono.merge(funded_scored_all[['unique_id', 'client', 'p1_overall_lda', 'p1_v1_5']], on= ['unique_id', 'client'], how= 'left')
print(oot_ono.shape)
print(oot_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].isnull().sum())

print(ttd_ono.shape)
ttd_ono= ttd_ono.merge(ttd_scored_all[['unique_id', 'client', 'p1_overall_lda', 'p1_v1_5']], on= ['unique_id', 'client'], how= 'left')
print(ttd_ono.shape)
print(ttd_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].isnull().sum())

ttd_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].describe(), train_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].describe(), test_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].describe(), oot_ono[['p1', 'p1_overall_lda', 'p1_v1_5']].describe()

train_ono.groupby(['dataset'])[['p1', 'p1_overall_lda', 'p1_v1_5']].apply(lambda x: x.describe()).T

train_ono[train_ono['dataset'].isin(['Property Owner'])].shape

train_ono['dataset'].value_counts(dropna= False)

def get_model_comparison_segment(train_d, test_d, oot_d, ttd_d, segment, pref, d):
    train_d= train_d.copy()
    test_d= test_d.copy()
    oot_d= oot_d.copy()
    ttd_d= ttd_d.copy()
    print({'segment': segment})#, 'overall shape': data.shape, 'segmented shape': data[data['dataset'].isin(segment)].shape})
    
    a= {'Model': pref, 'Dataset': d}
    
    a[f'KS Weighted Train Model']= rnkOrderingN(train_d[train_d['dataset'].isin(segment)], 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Weighted Test Model']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS']
    
    a[f'KS Non Weighted Train Model']= rnkOrderingN(train_d[train_d['dataset'].isin(segment)], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Non Weighted Test Model']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Non Weighted Oot Model']= rnkOrderingN(oot_d[oot_d['dataset'].isin(segment)], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS']
    
    a[f'KS Non Weighted Oot FICO']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'fico_pb', n_bkts= 5, asc= True)[1]['KS']
    a[f'KS Non Weighted Oot Auto Score']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'auto_score', n_bkts= 5, asc= True)[1]['KS']
    
    a['Bad Rate Train']= train_d[train_d['dataset'].isin(segment)]['bad_flag'].mean()
    a['Bad Rate Test']= test_d[test_d['dataset'].isin(segment)]['bad_flag'].mean()
    a['Bad Rate Oot']= oot_d[oot_d['dataset'].isin(segment)]['bad_flag'].mean()
    
    air_sgement_model= air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_d[ttd_d['dataset'].isin(segment)], 'approve_flag', 'p1'), 'AIR TTD')
    a[f'AIR Age Model']= air_sgement_model['AIR TTDage_cuts']
    a[f'AIR Race Model']= air_sgement_model['AIR TTDRace']
    a[f'AIR Race Black Model']= air_sgement_model['AIR TTDRaceblack']
    a[f'AIR Race Hispanic Model']= air_sgement_model['AIR TTDRacehispanic']
    a[f'AIR Gender Model']= air_sgement_model['AIR TTDGender']
    
    air_fico_model= air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_d[ttd_d['dataset'].isin(segment)], 'approve_flag', 'fico_pb', asc= False), 'AIR TTD')
    a[f'AIR Age FICO']= air_fico_model['AIR TTDage_cuts']
    a[f'AIR Race FICO']= air_fico_model['AIR TTDRace']
    a[f'AIR Race Black FICO']= air_fico_model['AIR TTDRaceblack']
    a[f'AIR Race Hispanic FICO']= air_fico_model['AIR TTDRacehispanic']
    a[f'AIR Gender FICO']= air_fico_model['AIR TTDGender']
    
    return a

def get_model_comparison_overall(train_d, test_d, oot_d, ttd_d, segment, pref, d):
    train_d= train_d.copy()
    test_d= test_d.copy()
    oot_d= oot_d.copy()
    ttd_d= ttd_d.copy()
    print({'segment': segment})#, 'overall shape': data.shape, 'segmented shape': data[data['dataset'].isin(segment)].shape})
    
    a= {'Model': pref, 'Dataset': d}
    
    a[f'KS Weighted Train Model']= rnkOrderingN(train_d[train_d['dataset'].isin(segment)], 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Weighted Test Model']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS']
    
    a[f'KS Non Weighted Train Model']= rnkOrderingN(train_d[train_d['dataset'].isin(segment)], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Non Weighted Test Model']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Non Weighted Oot Model']= rnkOrderingN(oot_d[oot_d['dataset'].isin(segment)], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS']
    
    a[f'KS Non Weighted Oot FICO']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'fico_pb', n_bkts= 5, asc= True)[1]['KS']
    a[f'KS Non Weighted Oot Auto Score']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'auto_score', n_bkts= 5, asc= True)[1]['KS']
    
    a['Bad Rate Train']= train_d[train_d['dataset'].isin(segment)]['bad_flag'].mean()
    a['Bad Rate Test']= test_d[test_d['dataset'].isin(segment)]['bad_flag'].mean()
    a['Bad Rate Oot']= oot_d[oot_d['dataset'].isin(segment)]['bad_flag'].mean()
    
    air_sgement_model= air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_d[ttd_d['dataset'].isin(segment)], 'approve_flag', 'p1_overall_lda'), 'AIR TTD')
    a[f'AIR Age Model']= air_sgement_model['AIR TTDage_cuts']
    a[f'AIR Race Model']= air_sgement_model['AIR TTDRace']
    a[f'AIR Race Black Model']= air_sgement_model['AIR TTDRaceblack']
    a[f'AIR Race Hispanic Model']= air_sgement_model['AIR TTDRacehispanic']
    a[f'AIR Gender Model']= air_sgement_model['AIR TTDGender']
    
    air_fico_model= air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_d[ttd_d['dataset'].isin(segment)], 'approve_flag', 'fico_pb', asc= False), 'AIR TTD')
    a[f'AIR Age FICO']= air_fico_model['AIR TTDage_cuts']
    a[f'AIR Race FICO']= air_fico_model['AIR TTDRace']
    a[f'AIR Race Black FICO']= air_fico_model['AIR TTDRaceblack']
    a[f'AIR Race Hispanic FICO']= air_fico_model['AIR TTDRacehispanic']
    a[f'AIR Gender FICO']= air_fico_model['AIR TTDGender']
    
    return a
    

def get_model_comparison_v15(train_d, test_d, oot_d, ttd_d, segment, pref, d):
    train_d= train_d.copy()
    test_d= test_d.copy()
    oot_d= oot_d.copy()
    ttd_d= ttd_d.copy()
    print({'segment': segment})#, 'overall shape': data.shape, 'segmented shape': data[data['dataset'].isin(segment)].shape})
    
    a= {'Model': pref, 'Dataset': d}
    
    a[f'KS Weighted Train Model']= rnkOrderingN(train_d[train_d['dataset'].isin(segment)], 'bad_flag', 'wt2', 'p1_v1_5', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Weighted Test Model']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', 'wt2', 'p1_v1_5', n_bkts= 5, asc= False)[1]['KS']
    
    a[f'KS Non Weighted Train Model']= rnkOrderingN(train_d[train_d['dataset'].isin(segment)], 'bad_flag', None, 'p1_v1_5', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Non Weighted Test Model']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'p1_v1_5', n_bkts= 5, asc= False)[1]['KS']
    a[f'KS Non Weighted Oot Model']= rnkOrderingN(oot_d[oot_d['dataset'].isin(segment)], 'bad_flag', None, 'p1_v1_5', n_bkts= 5, asc= False)[1]['KS']
    
    a[f'KS Non Weighted Oot FICO']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'fico_pb', n_bkts= 5, asc= True)[1]['KS']
    a[f'KS Non Weighted Oot Auto Score']= rnkOrderingN(test_d[test_d['dataset'].isin(segment)], 'bad_flag', None, 'auto_score', n_bkts= 5, asc= True)[1]['KS']
    
    a['Bad Rate Train']= train_d[train_d['dataset'].isin(segment)]['bad_flag'].mean()
    a['Bad Rate Test']= test_d[test_d['dataset'].isin(segment)]['bad_flag'].mean()
    a['Bad Rate Oot']= oot_d[oot_d['dataset'].isin(segment)]['bad_flag'].mean()
    
    air_sgement_model= air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_d[ttd_d['dataset'].isin(segment)], 'approve_flag', 'p1_v1_5'), 'AIR TTD')
    a[f'AIR Age Model']= air_sgement_model['AIR TTDage_cuts']
    a[f'AIR Race Model']= air_sgement_model['AIR TTDRace']
    a[f'AIR Race Black Model']= air_sgement_model['AIR TTDRaceblack']
    a[f'AIR Race Hispanic Model']= air_sgement_model['AIR TTDRacehispanic']
    a[f'AIR Gender Model']= air_sgement_model['AIR TTDGender']
    
    air_fico_model= air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_d[ttd_d['dataset'].isin(segment)], 'approve_flag', 'fico_pb', asc= False), 'AIR TTD')
    a[f'AIR Age FICO']= air_fico_model['AIR TTDage_cuts']
    a[f'AIR Race FICO']= air_fico_model['AIR TTDRace']
    a[f'AIR Race Black FICO']= air_fico_model['AIR TTDRaceblack']
    a[f'AIR Race Hispanic FICO']= air_fico_model['AIR TTDRacehispanic']
    a[f'AIR Gender FICO']= air_fico_model['AIR TTDGender']
    
    return a

pd.DataFrame(air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono, 'approve_flag', 'fico_pb', asc= False), 'AIR TTD'))

pd.DataFrame(air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Property Owner'], 'approve_flag', 'fico_pb', asc= False), 'AIR TTD'))

pd.DataFrame(air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Non Property Owner'], 'approve_flag', 'fico_pb', asc= False), 'AIR TTD'))

pd.concat(
    [pd.DataFrame(get_model_comparison_segment(train_ono, test_ono, oot_ono, ttd_ono, ['Property Owner'], 'Property Owner', 'Property Owner')),
     pd.DataFrame(get_model_comparison_overall(train_ono, test_ono, oot_ono, ttd_ono, ['Property Owner'], 'Overall', 'Property Owner')),
     pd.DataFrame(get_model_comparison_v15(train_ono, test_ono, oot_ono, ttd_ono, ['Property Owner'], 'V1.5', 'Property Owner'))]
)

pd.concat(
    [pd.DataFrame(get_model_comparison_segment(train_ono, test_ono, oot_ono, ttd_ono, ['Non Property Owner'], 'Non Property Owner', 'Non Property Owner')),
     pd.DataFrame(get_model_comparison_overall(train_ono, test_ono, oot_ono, ttd_ono, ['Non Property Owner'], 'Overall', 'Non Property Owner')),
     pd.DataFrame(get_model_comparison_v15(train_ono, test_ono, oot_ono, ttd_ono, ['Non Property Owner'], 'V1.5', 'Non Property Owner'))]
)

pd.concat(
    [pd.DataFrame(get_model_comparison_segment(train_ono, test_ono, oot_ono, ttd_ono, ['Non Property Owner', 'Property Owner'], 'Both', 'Both')),
     pd.DataFrame(get_model_comparison_overall(train_ono, test_ono, oot_ono, ttd_ono, ['Non Property Owner', 'Property Owner'], 'Overall', 'Both')),
     pd.DataFrame(get_model_comparison_v15(train_ono, test_ono, oot_ono, ttd_ono, ['Non Property Owner', 'Property Owner'], 'V1.5', 'Both'))]
)

### END

def get_results(data, data_ttd, target, score, cols_tuple, weight, dataset, asc= False):
    pred= data.copy()
    a= {}
    
    a['dataset']= dataset
    a['model']= score
    a['weight']= weight
    
    t= get_risk_rank(data, target, cols_tuple, score, weight, asc)
    
    a['KS'], a['GINI']= t[1]['KS'], t[1]['Gini']
    
    pred = Equal_Approval_Cutoff_Barsath(data_ttd, 'approve_flag', score)
    air_ttd = air_calc_new(pred,'AIR ttd ')
    a.update({k: v[0] for k, v in air_ttd.items()})
    # print(a)
    
    return a

# get_results(train_ono[train_ono['dataset']== 'Property Owner'], ttd_ono[ttd_ono['dataset']== 'Property Owner'], 'bad_flag', 'p1', col_tuple_pass, 'conwt')

test_ono.head()

print('Property Owner')
print('Bad Rate Train', train_ono[train_ono['dataset']== 'Property Owner']['bad_flag'].mean())
print('Bad Rate Test', test_ono[test_ono['dataset']== 'Property Owner']['bad_flag'].mean())
print('Bad Rate Oot', oot_ono[oot_ono['dataset']== 'Property Owner']['bad_flag'].mean())

print('KS Weighted Train p1', rnkOrderingN(train_ono[train_ono['dataset']== 'Property Owner'], 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Weighted Test p1', rnkOrderingN(test_ono[test_ono['dataset']== 'Property Owner'], 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS'])
      
print('KS Non Weighted Train p1', rnkOrderingN(train_ono[train_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Test p1', rnkOrderingN(test_ono[test_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Oot p1', rnkOrderingN(oot_ono[oot_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])

print('KS Non Weighted Test FICO', rnkOrderingN(test_ono[test_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'fico_pb', n_bkts= 5, asc= True)[1]['KS'])
print('KS Non Weighted Test Auto Score', rnkOrderingN(test_ono[test_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'auto_score', n_bkts= 5, asc= True)[1]['KS'])

print('AIR Table', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Property Owner'], 'approve_flag', 'p1'), 'AIR TTD'))

print('KS Weighted Train p1 overall', rnkOrderingN(train_ono[train_ono['dataset']== 'Property Owner'], 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Weighted Test p1 overall', rnkOrderingN(test_ono[test_ono['dataset']== 'Property Owner'], 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
      
print('KS Non Weighted Train p1 overall', rnkOrderingN(train_ono[train_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Test p1 overall', rnkOrderingN(test_ono[test_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Oot p1 overall', rnkOrderingN(oot_ono[oot_ono['dataset']== 'Property Owner'], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])

print('AIR Table', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Property Owner'], 'approve_flag', 'p1_overall_lda'), 'AIR TTD'))
print('AIR Table FICO', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Property Owner'], 'approve_flag', 'fico_pb', False), 'AIR TTD'))

print('Non Property Owner')
print('Bad Rate Train', train_ono[train_ono['dataset']== 'Non Property Owner']['bad_flag'].mean())
print('Bad Rate Test', test_ono[test_ono['dataset']== 'Non Property Owner']['bad_flag'].mean())
print('Bad Rate Oot', oot_ono[oot_ono['dataset']== 'Non Property Owner']['bad_flag'].mean())

print('KS Weighted Train p1', rnkOrderingN(train_ono[train_ono['dataset']== 'Non Property Owner'], 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Weighted Test p1', rnkOrderingN(test_ono[test_ono['dataset']== 'Non Property Owner'], 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS'])
      
print('KS Non Weighted Train p1', rnkOrderingN(train_ono[train_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Test p1', rnkOrderingN(test_ono[test_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Oot p1', rnkOrderingN(oot_ono[oot_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])

print('KS Non Weighted Test FICO', rnkOrderingN(test_ono[test_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'fico_pb', n_bkts= 5, asc= True)[1]['KS'])
print('KS Non Weighted Test Auto Score', rnkOrderingN(test_ono[test_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'auto_score', n_bkts= 5, asc= True)[1]['KS'])

print('AIR Table', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Non Property Owner'], 'approve_flag', 'p1'), 'AIR TTD'))

print('KS Weighted Train p1 overall', rnkOrderingN(train_ono[train_ono['dataset']== 'Non Property Owner'], 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Weighted Test p1 overall', rnkOrderingN(test_ono[test_ono['dataset']== 'Non Property Owner'], 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
      
print('KS Non Weighted Train p1 overall', rnkOrderingN(train_ono[train_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Test p1 overall', rnkOrderingN(test_ono[test_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Oot p1 overall', rnkOrderingN(oot_ono[oot_ono['dataset']== 'Non Property Owner'], 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])

print('AIR Table', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Non Property Owner'], 'approve_flag', 'p1_overall_lda'), 'AIR TTD'))
print('AIR Table FICO', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono[ttd_ono['dataset']== 'Non Property Owner'], 'approve_flag', 'fico_pb', False), 'AIR TTD'))

train_ono['dataset'].value_counts(dropna= False)

print('Overall')
print('Bad Rate Train', train_ono['bad_flag'].mean())
print('Bad Rate Test', test_ono['bad_flag'].mean())
print('Bad Rate Oot', oot_ono['bad_flag'].mean())

print('KS Weighted Train p1', rnkOrderingN(train_ono, 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Weighted Test p1', rnkOrderingN(test_ono, 'bad_flag', 'wt2', 'p1', n_bkts= 5, asc= False)[1]['KS'])
      
print('KS Non Weighted Train p1', rnkOrderingN(train_ono, 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Test p1', rnkOrderingN(test_ono, 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Oot p1', rnkOrderingN(oot_ono, 'bad_flag', None, 'p1', n_bkts= 5, asc= False)[1]['KS'])

print('KS Non Weighted Test FICO', rnkOrderingN(test_ono, 'bad_flag', None, 'fico_pb', n_bkts= 5, asc= True)[1]['KS'])
print('KS Non Weighted Test Auto Score', rnkOrderingN(test_ono, 'bad_flag', None, 'auto_score', n_bkts= 5, asc= True)[1]['KS'])

print('AIR Table', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono, 'approve_flag', 'p1'), 'AIR TTD'))

print('KS Weighted Train p1 overall', rnkOrderingN(train_ono, 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Weighted Test p1 overall', rnkOrderingN(test_ono, 'bad_flag', 'wt2', 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
      
print('KS Non Weighted Train p1 overall', rnkOrderingN(train_ono, 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Test p1 overall', rnkOrderingN(test_ono, 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])
print('KS Non Weighted Oot p1 overall', rnkOrderingN(oot_ono, 'bad_flag', None, 'p1_overall_lda', n_bkts= 5, asc= False)[1]['KS'])

print('AIR Table', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono, 'approve_flag', 'p1_overall_lda'), 'AIR TTD'))
print('AIR Table FICO', air_calc_new(Equal_Approval_Cutoff_Barsath(ttd_ono, 'approve_flag', 'fico_pb', False), 'AIR TTD'))

pw= train_ono[train_ono['dataset']== 'Property Owner'].reset_index(drop= True)
pw.shape

npw= train_ono[train_ono['dataset']== 'Non Property Owner'].reset_index(drop= True)
npw.shape

pw['bad_flag'].mean()*100, npw['bad_flag'].mean()*100, npw['bad_flag'].sum(), pw['bad_flag'].sum()

pw_good= pw[pw['bad_flag']!= 1]
pw_bad= pw[pw['bad_flag']== 1]

print(pw_good.shape, pw_bad.shape)

pw_good= pw_good.sample(frac= 0.5, random_state= 101)
print(pw_good.shape)

pw= pd.concat([pw_good, pw_bad]).reset_index(drop= True)
print(pw.shape, pw['bad_flag'].sum(), pw['bad_flag'].mean())

npw_good= npw[npw['bad_flag']!= 1]
npw_bad= npw[npw['bad_flag']== 1]

print(npw_good.shape, npw_bad.shape)

npw_good= npw_good.sample(frac= 0.85, random_state= 101)
print(npw_good.shape)

npw_bad= npw_bad.sample(frac= 0.38, random_state= 101)
print(npw_bad.shape)

npw= pd.concat([npw_good, npw_bad]).reset_index(drop= True)
print(npw.shape, npw['bad_flag'].sum(), npw['bad_flag'].mean())

print(train_ono.shape, train_ono['bad_flag'].sum(), train_ono['bad_flag'].mean())
train_ono= pd.concat([pw, npw]).reset_index(drop= True)
print(train_ono.shape, train_ono['bad_flag'].sum(), train_ono['bad_flag'].mean())
