==> Auto_Codes/9.1_Auto_Moving_Data_to_S3.ipynb <==
!pip install orjson

import warnings
warnings.filterwarnings('ignore')

import importlib
import os
import boto3
import re
import pandas as pd
import orjson

# Specify the path to your module
module_path = '/root/Unified_LN_V2/DataScience-Repo/SageMaker/read_data_in_sagemaker.py'
#module_path = '/home/ec2-user/SageMaker/DataScience-Repo/SageMaker/read_data_in_sagemaker.py'

# Create a spec for the module
module_spec = importlib.util.spec_from_file_location('read_data_in_sagemaker', module_path)

# Import the module
read_data_in_sagemaker = importlib.util.module_from_spec(module_spec)
module_spec.loader.exec_module(read_data_in_sagemaker)

s3 = boto3.client(service_name="s3",region_name = "us-east-2")

pd.set_option('display.max_columns', 500)
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_colwidth', 500)

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/')

!ls '/root/Unified_LN_V2/LN_Auto_V2/Data/'

s3_resource = boto3.resource('s3')

#s3_resource.Bucket(bucket_name).upload_file(source_file_path_in_sagemaker, destination_file_path_in_s3_bucket)
s3_resource.Bucket('scien-ds-sm-root-backup').upload_file("/root/Unified_LN_V2/LN_Auto_V2/Data/2_LN_Auto_V2_Train_Funded_Data_250324.csv", "home/sushil/LN_Unified_V2/Auto_Model_Dataset/2_LN_Auto_V2_Train_Funded_Data_250324.csv")

for file in ['2_LN_Auto_V2_Test_Funded_Data_250324.csv', '2_LN_Auto_V2_TTD_Data_250324.csv',
             '2_LN_Auto_V2_Oot_Funded_Data_250324.csv']:
    s3_resource.Bucket('scien-ds-sm-root-backup').upload_file(
        f"/root/Unified_LN_V2/LN_Auto_V2/Data/{file}", f"home/sushil/LN_Unified_V2/Auto_Model_Dataset/Raw_Data/{file}")

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/Raw_Data/')

s3_resource.Bucket('scien-ds-sm-root-backup').upload_file(
    f"/root/Unified_LN_V2/LN_Auto_V2/Data/040424_Auto_LN_V1_5_Scored_Data.csv", f"home/sushil/LN_Unified_V2/Auto_Model_Dataset/Raw_Data/040424_Auto_LN_V1_5_Scored_Data.csv")

s3_resource.Bucket('scien-ds-sm-root-backup').upload_file(
    f"/root/Unified_LN_V2/LN_Auto_V2/Data/1_LN_Auto_All_Updated_Data_220324.csv", f"home/sushil/LN_Unified_V2/Auto_Model_Dataset/Raw_Data/1_LN_Auto_All_Updated_Data_220324.csv")

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/Raw_Data/')

s3_resource.Bucket('scien-ds-sm-root-backup').upload_file(
    f"/root/Unified_LN_V2/LN_Auto_V2/Model_Data/RI_Auto_Model/250324_RI_Auto_Model_Scored_Data.csv", f"home/sushil/LN_Unified_V2/Auto_Model_Dataset/RI/250324_RI_Auto_Model_Scored_Data.csv")

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/RI/')

!ls '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Auto_FICO_660_740_Model/'

!ls '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/'

os.listdir('/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Auto_FICO_660_740_Model/')

os.path.isdir('/root/Unified_LN_V2/LN_Auto_V2')

path= '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Auto_FICO_660_740_Model/'

s3_resource.Bucket('scien-ds-sm-root-backup').upload_file(
    f"/root/Unified_LN_V2/LN_Auto_V2/Model_Data/RI_Auto_Model/250324_RI_Auto_Model_Scored_Data.csv", f"home/sushil/LN_Unified_V2/Auto_Model_Dataset/RI/250324_RI_Auto_Model_Scored_Data.csv")

for i,j,k in os.walk(path):
    print(f'path- {i}')
    print(f'folders- {j}')
    print(f'files - {k}')
    # for file in k:
        # print(file)

for i, j, k in os.walk(path):
    print(i, j, k)

def upload_to_s3(path, path_to_upload, bucket= 'scien-ds-sm-root-backup'):
    import boto3
    pref= path.split('/')[-2]
    print(pref)
    s3_resource = boto3.resource('s3')
    
    for i, j, k in os.walk(path):
        for file in k:
            s3_resource.Bucket(bucket).upload_file(
                f"{i}/{file}", f"{path_to_upload}{pref}/{i.split('/')[-1]}/{file}"
            )

upload_to_s3(
    '/root/Unified_LN_V2/LN_Auto_V2/Model_Data/Without_LDA_Auto_LN_V2/', 
    'home/sushil/LN_Unified_V2/Auto_Model_Dataset/'
)

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/')

read_data_in_sagemaker.list_folders(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/Without_LDA_Auto_LN_V2/')

read_data_in_sagemaker.list_folders(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/')

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Auto_Model_Dataset/Without_LDA_Auto_LN_V2/Iteration_3/')

### Unsec

s3_resource.Bucket('scien-ds-sm-root-backup').upload_file(
    f"/root/Unified_LN_V2/LN_Unsec_V2/Data/1_LN_Unsec_All_Updated_Data_090424.csv", f"home/sushil/LN_Unified_V2/Unsec_Model_Dataset/1_LN_Unsec_All_Updated_Data_090424.csv")

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Unsec_Model_Dataset/')

read_data_in_sagemaker.list_folders(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/')

os.listdir('/root/Unified_LN_V2/LN_Unsec_V2/Model_Data/')

upload_to_s3(
    '/root/Unified_LN_V2/LN_Unsec_V2/Model_Data/Protected_Class_Data/', 
    'home/sushil/LN_Unified_V2/Unsec_Model_Dataset/'
)

read_data_in_sagemaker.list_folders(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Unsec_Model_Dataset/Without_Segment_Model/')

read_data_in_sagemaker.list_objects_in_folder(s3, 'scien-ds-sm-root-backup', 'home/sushil/LN_Unified_V2/Unsec_Model_Dataset')
